Add better DOCTYPE parsing
This change adds the `PatentXmlToTabular.xml_root` property which can be used to check the XML doctype and filtering when parsing files that contain multiple XML docs that are not all the target type (as with the USPTO data).

The `xml_root` can be specified in the YAML config file (as `<xml_root>: ...` at the top level of the config file -- if it is not specified in this fashion, the first encountered path is assumed to be the root element (the assumption here being that most config files will be structured this way).

Closes #11.