Instantiate etree.XMLParser on the doc parser
`etree.XMLParser` is implemented in C and so can't be picked (even by dill).  This is a problem because all `multiprocessing.pool` methods use `multiprocessing.SimpleQueue` to pass tasks to the worker processes, and everything that goes through `multiprocessing.SimpleQueue` must be picklable.

So in preparation for building in mp support, we must instead pass the needed arguments and construct the `XMLParser` in the child processes.